Moved box in separate file
